Refactor round #2
